fixed the PR comments + code cleanup
