+++ src/main/java/com/bootcamp/demo/restapi/FirebaseController.java
@@ -1,8 +1,7 @@
-package com.bootcamp.demo.firebase;
+package com.bootcamp.demo.restapi;
 
 import com.google.cloud.firestore.CollectionReference;
 import com.google.cloud.firestore.Firestore;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -17,7 +16,6 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
  * Sample RestController
  * Demo - Used for testing purposes
  */
-@ConditionalOnProperty("firebaseKey")
 @RestController
 @RequestMapping(path = "/firebase", produces = APPLICATION_JSON_VALUE)
 public class FirebaseController {
