+++ src/main/java/com/bootcamp/demo/service/exception/ServiceException.java
@@ -2,7 +2,15 @@ package com.bootcamp.demo.service.exception;
 
 public class ServiceException extends RuntimeException {
 
-    public ServiceException() {
-        super("");
+    public ServiceException(String message) {
+        super(message);
+    }
+
+    public ServiceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ServiceException(Throwable cause) {
+        super(cause);
     }
 }
